+++ custom_components/lernende_heizung/const.py
@@ -6,9 +6,9 @@ from typing import Final
 
 DOMAIN: Final = "lernende_heizung"
 NAME: Final = "Lernende Heizung"
-VERSION: Final = "0.4.0"
+VERSION: Final = "0.5.0"
 
-PLATFORMS: Final = ["binary_sensor", "climate", "datetime", "select", "sensor", "switch"]
+PLATFORMS: Final = ["binary_sensor", "button", "climate", "datetime", "select", "sensor", "switch"]
 
 # Allgemeine Optionen
 CONF_WEATHER: Final = "weather_entity"